+++ app/routes/auth.callback.tsx
@@ -7,7 +7,7 @@ import {
   destroySession,
   getSession,
 } from "~/services/prototype.session.server";
-import { LoginError } from "./action.login";
+import { LoginError } from "./action.login-user";
 
 export const loader: LoaderFunction = async ({ request }) => {
   // Fyi: When the authorization server redirects to this route (Redirect URI),
