+++ tests/e2e/index.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "@playwright/test";
-import { LoginError } from "~/routes/action.login";
+import { LoginError } from "~/routes/action.login-user";
 import { LogoutType } from "~/routes/action.logout-user";
 
 test.describe("Homepage (_index route)", () => {
